feat(sidebar): move FY switcher and New FY modal into Sidebar.tsx (closes #228)
